requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cloudinary==1.30.0
dj-database-url==1.0.0
Django==4.1.3
django-autoslug==1.9.8
django-cities-light==3.9.2
gunicorn==20.1.0
Pillow==9.3.0
progressbar2==4.2.0
python-dotenv==0.21.0
python-utils==3.4.5
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
Unidecode==1.3.6
urllib3==1.26.12
